requirements.txt
matplotlib
djangorestframework
markdown
django-filter
numpy
django=3.2.9